+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 2.5.4
+
+* [core][antibodies] fix acowner table definition and fix delete javascript function #638
+
 ## 2.5.3
 
 * [invoice][services] fix services invoices (invoice id not set) #637